ServiceOrie_Sys/requirements.txt
streamlit
audiorecorder
python-dotenv
groq
elevenlabs
